fix: resolve clippy warnings
Fix all clippy lints flagged by CI:
- Collapse nested if-let statements in dwarf_variables.rs and solidity.rs (collapsible_if)
- Replace while loop with unconditional break with single if-let in solidity.rs (never_loop)
- Remove unnecessary u32-to-u32 cast in tracer.rs (unnecessary_cast)

+++ src/tracer.rs
@@ -169,7 +169,7 @@ impl PolkaVmTracer {
                     // Try to resolve source location from debug info.
                     let (source_path, line) = source_mapper
                         .resolve(pc)
-                        .unwrap_or_else(|| (blob_path, pc.0 as u32 + 1));
+                        .unwrap_or_else(|| (blob_path, pc.0 + 1));
 
                     // Emit step if line changed.
                     if prev_line != Some(line) {
